part2 done but the color of the normals is weird

+++ part2/source/utils.cpp
@@ -1,6 +1,6 @@
 
 #include "utils.h"
 
-int map(int value, int istart, int istop, int ostart, int ostop) {
+float map(float value, float istart, float istop, float ostart, float ostop) {
     return ostart + (ostop - ostart) * (((value - istart)*1.0) / ((istop - istart)-1.0));
 }
